lib: rework custom export code to use environment file exclusively
This allows the behavior to be easily replicated between a local machine
and the CI, while allowing the simplification of other code.

Also add the hash of lib.sh to the cache key, which will let us test if
changes to the support library break any portions of the pipeline.