nightlies: add build revisioning system [deploy]
This adds versioning to nightlies build for a specific commit + branch
combination. This allows the publication of newer nightly builds
for the same commit + branch combination to be incrementally versioned.

This is a breaking change for all current nightlies user.

Tags created by nightlies will now use the following format:

    <date>-<branch>-<commit>-<revision>

In addition, archives published by nightlies will now use the following format:

    <nim>-<version>-<revision><suffix>.<ext>

The main change is essentially the addition of a <revision> field. This field
is local to the commit + branch combination of the complier source, and is
incremented whenever nightlies received a revision bump which requires the
rebuild of current assets.